+++ ui/packages/shared/profile/src/ProfileIcicleGraph/IcicleGraph/utils.ts
@@ -31,6 +31,9 @@ export const getBinaryName = (
   if (node.meta?.locationIndex === undefined || node.meta?.locationIndex === 0) {
     return undefined;
   }
+  if (node.meta.locationIndex > locations.length) {
+    return undefined;
+  }
 
   const location = locations[node.meta.locationIndex - 1];
 
@@ -57,7 +60,13 @@ export function nodeLabel(
   if (node.meta?.locationIndex === undefined) return '<unknown>';
   if (node.meta?.locationIndex === 0) return '<unknown>';
 
+  if (node.meta.locationIndex > locations.length) {
+    console.info('location index out of bounds', node.meta.locationIndex, locations.length);
+    return '<unknown>';
+  }
+
   const location = locations[node.meta.locationIndex - 1];
+  if (location === undefined) return '<unknown>';
 
   let mappingString = '';